requirements.txt
nextcord
hmtai
pyyaml